+++ test/cambium/core_test.clj
@@ -56,6 +56,10 @@
           (is (= "quux" (c/context-val :baz)) "Delta context override must not remove non-overridden attributes")
           (is (= "baz" (c/context-val :bar)))))
       (is (nil? (c/context-val :foo)) "Attribute not set must be absent after restoration"))
+    (testing "deletion via nil values"
+      (c/with-logging-context context-old
+        (c/with-logging-context {:foo nil}
+          (is (not (contains? (c/get-context) (c/stringify-key :foo)))))))
     (testing "wrap-raw-mdc"
       (is (nil? (c/context-val :foo)))
       ((c/wrap-logging-context context-old f))